Extract ClearSelection listener
