+++ frontend/pages/_app.tsx
@@ -5,6 +5,7 @@ import axios from "axios";
 import axiosCaseConverter from "axios-case-converter";
 import { RecoilRoot } from "recoil";
 import Errors from "components/Errors";
+import ClearSelection from "listeners/ClearSelection";
 
 const API = axiosCaseConverter(
   axios.create({
@@ -16,6 +17,7 @@ const App: FC<AppProps> = ({ Component, pageProps }) => {
   return (
     <RecoilRoot>
       <Errors />
+      <ClearSelection />
       <Component {...pageProps} />
     </RecoilRoot>
   );
